pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netops-skills"
version = "0.1.0"
description = "Network Operations Skills - Minimize repetitive typing for NOC tasks"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "questionary>=2.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
netops = "netops_skills.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["netops_skills*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
